\name{c212.FDR.data}
\docType{data}
\alias{c212.FDR.data}
\title{Fisher Test p-values for End of Trial Data Clinical Data}
\description{
  This data set the p-values for a comparison between adverse events on each arm of a clinical trial.
}
\usage{c212.FDR.data}
\format{A dataframe with columns \emph{B} - body-system, \emph{AE} - adverse event, \emph{p} - p-value for two-sided Fisher exact test.
The dataframe contains 45 observations.}
%\source{World Almanac and Book of Facts, 1975, page 406.}
%\references{
%  McNeil, D. R. (1977) \emph{Interactive Data Analysis}.
%  New York: Wiley.
%}
\keyword{datasets}
