\name{c212.trial.data}
\docType{data}
\alias{c212.trial.data}
\title{End of Trial Data Clinical Data for Adverse Event Incidence}
\description{
  This data set contains the counts of adverse event incidence for the trial.
}
\usage{data(c212.trial.data)}
\format{A dataframe with columns \emph{B} - body-system, \emph{AE} - adverse event, \emph{Group} - 1 for control, 2 for treatment, \emph{Count} - total adverse event incidence, \emph{Total} - total patients on the trial arm.
The dataframe contains 90 observations.}
%\source{World Almanac and Book of Facts, 1975, page 406.}
%\references{
%  McNeil, D. R. (1977) \emph{Interactive Data Analysis}.
%  New York: Wiley.
%}
\keyword{datasets}
