\name{c212.trial.interval.data1}
\docType{data}
\alias{c212.trial.interval.data1}
\title{Interim analysis trial data.}
\description{
  This data set contains count of the adverse events over the first two intervals of a clinical trial.
}
\usage{c212.trial.interval.data1}
\format{A dataframe with columns \emph{interval}, \emph{I_index} - the interval order, \emph{B} - body-system, \emph{AE} - adverse event, \emph{Group} - 1 for control, 2 for treatment, \emph{Count} - total adverse events that occurred in the interval, \emph{Exposure} - the total time all at risk subjects spent in the interval.
The dataframe contains 1860 observations.}
%\source{World Almanac and Book of Facts, 1975, page 406.}
%\references{
%  McNeil, D. R. (1977) \emph{Interactive Data Analysis}.
%  New York: Wiley.
%}
\keyword{datasets}
